// libs/web/features/github/model/github.types.ts
export interface RepoQueryParams {
  per_page?: number;
  sort?: "created" | "updated" | "pushed" | "full_name";
  direction?: "asc" | "desc";
}
// ...
export interface PullRequestQueryParams {
  state?: "open" | "closed" | "all";
  per_page?: number;
  sort?: "created" | "updated" | "popularity" | "long-running";
  direction?: "asc" | "desc";
}